requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=1.4.28,<2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0

# YouTube API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data processing
pandas==2.1.4
jinja2==3.1.2

# Scheduling
apache-airflow==2.7.3
schedule==1.2.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 